refactor: depend on external crate
