+++ validity/src/lib.rs
@@ -5,7 +5,7 @@ mod env;
 mod logger;
 mod prom;
 pub mod proof_requester;
-mod proofs;
+mod proofs_service;
 mod proposer;
 mod signer;
 mod types;
@@ -18,7 +18,7 @@ pub use env::*;
 pub use logger::*;
 pub use prom::*;
 pub use proof_requester::*;
-pub use proofs::*;
+pub use proofs_service::*;
 pub use proposer::*;
 pub use signer::*;
 pub use types::*;
